Decoupled components and added blank states for form builder
